+++ src/main/java/com/example/happyhousebackend/domain/family/repository/FamilyRepository.java
@@ -1,5 +1,6 @@
-package com.example.happyhousebackend.domain.family;
+package com.example.happyhousebackend.domain.family.repository;
 
+import com.example.happyhousebackend.domain.family.entity.Family;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FamilyRepository extends JpaRepository<Family, Long> {
