+++ src/main/java/com/example/happyhousebackend/domain/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.example.happyhousebackend.domain.member;
+package com.example.happyhousebackend.domain.member.repository;
 
+import com.example.happyhousebackend.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
