+++ src/main/java/com/example/happyhousebackend/domain/notification/repository/NotificationRepository.java
@@ -1,5 +1,6 @@
-package com.example.happyhousebackend.domain.notification;
+package com.example.happyhousebackend.domain.notification.repository;
 
+import com.example.happyhousebackend.domain.notification.entity.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
