--- src/main/java/com/example/happyhousebackend/domain/routine/RoutineCompletedRepository.java
@@ -1,6 +1,0 @@
-package com.example.happyhousebackend.domain.routine;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoutineCompletedRepository extends JpaRepository<RoutineCompleted, RoutineCompletedPK> {
-}
