+++ src/main/java/com/example/happyhousebackend/domain/routine/entity/SampleRoutine.java
@@ -1,4 +1,4 @@
-package com.example.happyhousebackend.domain.routine;
+package com.example.happyhousebackend.domain.routine.entity;
 
 import lombok.*;
