+++ src/main/java/com/example/happyhousebackend/domain/routine/repository/RoutineRepeatRepository.java
@@ -1,5 +1,6 @@
-package com.example.happyhousebackend.domain.routine;
+package com.example.happyhousebackend.domain.routine.repository;
 
+import com.example.happyhousebackend.domain.routine.entity.RoutineRepeat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoutineRepeatRepository extends JpaRepository<RoutineRepeat, Long> {
